PackageTracker/requirements.txt
Django==2.2.1
msgpack==0.6.1
paho-mqtt==1.4.0
pip==19.1.1
pytz==2019.1
setuptools==41.0.1
sqlparse==0.3.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
Pillow==6.0.0
